+++ mobile/lib/infrastructure/loaders/remote_image_request.dart
@@ -1,158 +1,4 @@
-import 'dart:async';
-import 'dart:ffi';
-import 'dart:io';
-import 'dart:ui' as ui;
-
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter/foundation.dart';
-import 'package:immich_mobile/domain/models/asset/base_asset.model.dart';
-import 'package:immich_mobile/providers/image/cache/remote_image_cache_manager.dart';
-import 'package:immich_mobile/providers/infrastructure/platform.provider.dart';
-import 'package:ffi/ffi.dart';
-import 'package:logging/logging.dart';
-
-abstract class ImageRequest {
-  static int _nextRequestId = 0;
-
-  final int requestId = _nextRequestId++;
-  bool _isCancelled = false;
-
-  get isCancelled => _isCancelled;
-
-  ImageRequest();
-
-  Future<ImageInfo?> load(ImageDecoderCallback decode, {double scale = 1.0});
-
-  void cancel() {
-    if (isCancelled) {
-      return;
-    }
-    _isCancelled = true;
-    return _onCancelled();
-  }
-
-  void _onCancelled();
-
-  Future<ui.FrameInfo?> _fromPlatformImage(Map<String, int> info) async {
-    final address = info['pointer'];
-    if (address == null) {
-      if (!kReleaseMode) {
-        debugPrint('Platform image request for $requestId was cancelled');
-      }
-      return null;
-    }
-
-    final pointer = Pointer<Uint8>.fromAddress(address);
-    try {
-      if (_isCancelled) {
-        return null;
-      }
-
-      final actualWidth = info['width']!;
-      final actualHeight = info['height']!;
-      final actualSize = actualWidth * actualHeight * 4;
-
-      final buffer = await ImmutableBuffer.fromUint8List(pointer.asTypedList(actualSize));
-      if (_isCancelled) {
-        return null;
-      }
-
-      final descriptor = ui.ImageDescriptor.raw(
-        buffer,
-        width: actualWidth,
-        height: actualHeight,
-        pixelFormat: ui.PixelFormat.rgba8888,
-      );
-      final codec = await descriptor.instantiateCodec();
-      if (_isCancelled) {
-        return null;
-      }
-
-      return await codec.getNextFrame();
-    } finally {
-      malloc.free(pointer);
-    }
-  }
-}
-
-class ThumbhashImageRequest extends ImageRequest {
-  final String thumbhash;
-
-  ThumbhashImageRequest({required this.thumbhash});
-
-  @override
-  Future<ImageInfo?> load(ImageDecoderCallback decode, {double scale = 1.0}) async {
-    if (_isCancelled) {
-      return null;
-    }
-
-    Stopwatch? stopwatch;
-    if (!kReleaseMode) {
-      stopwatch = Stopwatch()..start();
-    }
-    final Map<String, int> info = await thumbnailApi.getThumbhash(thumbhash);
-    if (!kReleaseMode) {
-      stopwatch!.stop();
-      debugPrint('Thumbhash request $requestId took ${stopwatch.elapsedMilliseconds}ms');
-    }
-    final frame = await _fromPlatformImage(info);
-    return frame == null ? null : ImageInfo(image: frame.image, scale: scale);
-  }
-
-  @override
-  void _onCancelled() {
-    if (!kReleaseMode) {
-      debugPrint('Thumbhash request $requestId for $thumbhash was cancelled');
-    }
-  }
-}
-
-class LocalImageRequest extends ImageRequest {
-  final String localId;
-  final int width;
-  final int height;
-  final AssetType assetType;
-
-  LocalImageRequest({required this.localId, required ui.Size size, required this.assetType})
-    : width = size.width.toInt(),
-      height = size.height.toInt();
-
-  @override
-  Future<ImageInfo?> load(ImageDecoderCallback decode, {double scale = 1.0}) async {
-    if (_isCancelled) {
-      return null;
-    }
-
-    Stopwatch? stopwatch;
-    if (!kReleaseMode) {
-      stopwatch = Stopwatch()..start();
-    }
-    final Map<String, int> info = await thumbnailApi.requestImage(
-      localId,
-      requestId: requestId,
-      width: width,
-      height: height,
-      isVideo: assetType == AssetType.video,
-    );
-    if (!kReleaseMode) {
-      stopwatch!.stop();
-      debugPrint(
-        'Local image request $requestId took ${stopwatch.elapsedMilliseconds}ms for $localId of size $width x $height',
-      );
-    }
-    final frame = await _fromPlatformImage(info);
-    return frame == null ? null : ImageInfo(image: frame.image, scale: scale);
-  }
-
-  @override
-  Future<void> _onCancelled() {
-    if (!kReleaseMode) {
-      debugPrint('Local image request $requestId for $localId of size $width x $height was cancelled');
-    }
-    return thumbnailApi.cancelImageRequest(requestId);
-  }
-}
+part of 'image_request.dart';
 
 class RemoteImageRequest extends ImageRequest {
   static final log = Logger('RemoteImageRequest');
